requirements.txt
amqp==5.2.0
asgiref==3.7.1
async-timeout==4.0.3
attrs==23.1.0
autopep8==2.3.1
billiard==4.2.0
Brotli==1.1.0
cairocffi==1.7.1
CairoSVG==2.7.1
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
django-currentuser==0.6.1
django-filter==23.2
django-phonenumber-field==7.1.0
django-redis==5.4.0
django-sequences==2.8
djangorestframework==3.14.0
djangorestframework-recursive==0.1.2
drf-spectacular==0.26.4
et-xmlfile==1.1.0
fonttools==4.53.1
gprof2dot==2024.6.6
html5lib==1.1
idna==3.4
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
kombu==5.3.7
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
phonenumbers==8.13.19
Pillow==10.0.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
pycodestyle==2.12.0
pycparser==2.22
pydyf==0.11.0
pyphen==0.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.6
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.0
six==1.16.0
sqlparse==0.4.4
tinycss2==1.3.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.7
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==52.5
webencodings==0.5.1
XlsxWriter==3.2.0
zopfli==0.2.3
